+++ apps/web/src/components/tx-flow/actions/ComboSubmit.tsx
@@ -4,17 +4,26 @@ import { Box } from '@mui/material'
 import WalletRejectionError from '@/components/tx/SignOrExecuteForm/WalletRejectionError'
 import ErrorMessage from '@/components/tx/ErrorMessage'
 import { TxFlowContext } from '../TxFlowProvider'
+import { useValidateTxData } from '@/hooks/useValidateTxData'
 
 const ComboSubmit = ({ onSubmit }: SlotComponentProps<SlotName.Submit>) => {
-  const { submitError, isRejectedByUser, setShouldExecute } = useContext(TxFlowContext)
+  const { txId, submitError, isRejectedByUser, setShouldExecute } = useContext(TxFlowContext)
   const slotItems = useSlot(SlotName.ComboSubmit)
   const [submitAction, setSubmitAction] = useState<string>('sign')
+
+  const [validationResult, , validationLoading] = useValidateTxData(txId)
+  const validationError = useMemo(
+    () => (validationResult !== undefined ? new Error(validationResult) : undefined),
+    [validationResult],
+  )
 
   useEffect(() => {
     setShouldExecute(submitAction === 'execute')
   }, [submitAction, setShouldExecute])
 
   const options = useMemo(() => slotItems.map(({ label, id }) => ({ label, id })), [slotItems])
+
+  const disabled = validationError !== undefined || validationLoading
 
   return (
     <>
@@ -30,12 +39,17 @@ const ComboSubmit = ({ onSubmit }: SlotComponentProps<SlotName.Submit>) => {
         </Box>
       )}
 
+      {validationError !== undefined && (
+        <ErrorMessage error={validationError}>Error validating transaction data</ErrorMessage>
+      )}
+
       <Slot
         name={SlotName.ComboSubmit}
         id={submitAction}
         onSubmit={onSubmit}
         options={options}
         onChange={setSubmitAction}
+        disabled={disabled}
       />
     </>
   )
