+++ apps/web/src/components/tx-flow/actions/Sign/index.tsx
@@ -13,7 +13,7 @@ export const SIGN_CHECKBOX_TOOLTIP = 'Review details and check the box to enable
 
 const CheckboxGuardedSignForm = withCheckboxGuard(SignForm, SIGN_CHECKBOX_LABEL, SIGN_CHECKBOX_TOOLTIP)
 
-export const Sign = ({ onSubmit, options, onChange }: SlotComponentProps<SlotName.ComboSubmit>) => {
+export const Sign = ({ onSubmit, options, onChange, disabled = false }: SlotComponentProps<SlotName.ComboSubmit>) => {
   const [checked, setChecked] = useState(false)
   const { safeTx, txOrigin } = useContext(SafeTxContext)
   const { txId, trackTxEvent, isSubmittable } = useContext(TxFlowContext)
@@ -33,7 +33,7 @@ export const Sign = ({ onSubmit, options, onChange }: SlotComponentProps<SlotNam
 
   return (
     <CheckboxGuardedSignForm
-      disableSubmit={!isSubmittable}
+      disableSubmit={!isSubmittable || disabled}
       origin={txOrigin}
       safeTx={safeTx}
       onSubmit={handleSubmit}
